Exclude docs/conf.py from linting.
Running `make lint` after initializing a fresh project complains about errors in `docs/conf.py`. Which is kinda awkward for the user as he didn't write any code yet and already has errors. As this file is autogenerated I think it is best to exclude it from linting rather than fixing the errors.